packages: Rewrite search to use xq-api
This replaces the package search with a primitive one through xq-api.
Due to a limitation of the API, the architecture list is a drop-down
instead of being presented in the table -- so, users have to
unfortunately know which architecture they want to search.

The architectures list is populated by the API, so if a new one is
added (e.g., PowerPC), then it should be picked up automatically with
an xq-api reload.

This also moves the jquery and bootstrap scripts into the page head.
There's a chance this could slow down page rendering a bit, but without
the ability to add scripts to the end of the body after jquery, the
voidsearch.js and others can't make use of jquery reliably. So, a small
hit that shouldn't be noticable after caching seems better than writing
a bunch of code to work around browser load order.